requirements.txt
# Flask and extensions
Flask==3.0.0
Flask-Cors==4.0.0

# Google Cloud
google-cloud-datastore==2.19.0
google-cloud-storage==2.13.0

# Auth0 and JWT
PyJWT==2.8.0
python-jose[cryptography]==3.3.0
requests==2.31.0

# Environment and utilities
python-dotenv==1.0.0
gunicorn==21.2.0

# Development dependencies (optional)
pytest==7.4.3
pytest-flask==1.3.0
pytest-cov==4.1.0